Initial Commit.
A very popular question among different MNCs, to find an equilibrium point in an array.
.
An Equilibrium Point in an array is the point from where the sum of all its left side elements and the sum of all its right side elements are equal. If an array has only one element then that element is only the Equilibrium Point in that array and if an array has only two elements then no such Equilibrium Point exists in that array.
.
The problem can be solved very easily using the Brute Force approach or the Naive approach, with the Time Complexity of O(n^2).
.
But here the catch is to solve the problem in the linear Time Complexity i.e; in O(n). This has been achieved using a small mathematical calculation, and the Problem has been coded in C++.