Program Code (Using Merge Sort Technique).
This is a very popular question among the FAANG companies, to count the number of Inversions in an array, the number of Inversions is the number of times the array needs to be changed or modified so that the array becomes sorted.
.
Time Complexity: O(NlogN)
.
As the Merge Sort Technique has been used, therefore the Time Complexity to find the no. of Inversions in an array is O(NlogN).
.
Where N is the number of elements in an array.