Initial Commit (Brute Force Approach).
In this program, we have been given two arrays and we have to sort the first array according to the other array and display the resultant array. After sorting with respect to the second array we have to append the remaining elements of the first array in a sorted manner.